Mediante policy agregada restriccion para que unicamente el usuario que crea la vacante pueda borrarla, esto para evitar un usuario mal intencionado que modifique el id de la vacante desde la vista

+++ app/Http/Livewire/MostrarVacantes.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Models\Vacante;
 use Livewire\Component;
 
@@ -17,10 +18,16 @@ class MostrarVacantes extends Component
     //     );
     // }
 
+    // ? Con esto permitimos el uso del policy de vacantes en el metodo de eliminar, así evitamos que una persona mal intencionada modifique el id de la vacante y borre otro
+    use AuthorizesRequests;
+
     protected $listeners = ['eliminarVacante'];
     // public function eliminarVacante($id) // También sirve originalmente, afortunadamente livewire soporta route model binding y se simplifica como está abajo
     public function eliminarVacante(Vacante $vacante)
     {
+        // Primero verificamos que si pertenece a dicho usuario
+        $this->authorize('delete', $vacante);
+        // Luego eliminamos
         $vacante->delete();
     }
